将Tokenizer独立为一个文件
